--- backend/src/main/java/coffeeshout/room/application/port/DataInitializable.java
@@ -1,8 +1,0 @@
-package coffeeshout.room.application.port;
-
-import java.io.IOException;
-
-public interface DataInitializable {
-
-    void init() throws IOException;
-}
